+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 * [#512](https://github.com/bbatsov/rubocop/issues/512) - Fix bug causing crash in AndOr auto-correction.
 * [#515](https://github.com/bbatsov/rubocop/issues/515) - Fix bug causing AlignParameters and AlignArray auto-correction to destroy code.
 * [#516](https://github.com/bbatsov/rubocop/issues/516) - Fix bug causing RedundantReturn auto-correction to produce invalid code.
-
+* [#527](https://github.com/bbatsov/rubocop/issues/527) - Handle `!=` expressions in `EvenOdd` cop
 
 ## 0.13.1 (19/09/2013)